ica-regions.tsv,ica-super-regions.tsv - alter ID schema slightly
So that none of these regions or super-regions can possibly be
confused with an ISO alpha-2 country identifier, prefix them all with
`I-`.  Add a special super-region `I`, which represents the global
region.

We want these to be able to co-exist in the same namespace with ISO
2-character country codes. i.e. The juristiction of an institution
should be assignable to be assigned with any of these IDs and still be
unambiguous, even when full URIs are not used (as they are likely to
be in CSV columns, for example).