vocabs/standard/ - convert .skos files to .ttl equivalents
Where not already done (countries, regions, super-regions, which I
want to leave as generated by my recent efforts). Except update
activities.ttl for consistency, because we're using rapper for
conversion, instead of rdfpipe (again for consistency with the build
script; we don't want to add another frivolous dependency).

The converted turtle is a bit verbose, but clean-ups can be done later.