vocabs/standard/*.ttl - round trip through TSV
...using ttl2phrasetab and phrasetab2vocab (which should probably now
be called phrasetab2ttl).

  for i in *.ttl ; do echo $i; b=${i%.ttl}; bundle exec ./ttl2phrasetab <$i | tee $b.tsv | bundle exec ./phrasetab2vocab >${b}2.ttl; bundle exec ./rdf-diff $i ${b}2.ttl > $b.diff; mv -v ${b}2.ttl $i; done

The results have been diffed using rdf-diff, and the differences are acceptable:

- extra lang tags in some places (where lang tags are used and they had differed)
- loss of all lang tags when they are all empty (just a plain string used instead)
- updating of the dc:modified property

The physical order and structure of the TTL files has changed quite a
lot, however this should now remain fairly static on future
round-trips, and differences will be due to changes in the phrasetabs.